Added Timer

+++ src/Game/Level.cpp
@@ -132,7 +132,7 @@ void Level::render() const
 	}
 }
 
-void Level::update(const uint64_t delta)
+void Level::update(const double delta)
 {
 	for (const auto& currentLevelObject : m_levelObjects)
 	{
